pyproject.toml
[tool.poetry]
name = "meeting-script"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
PyGithub = "^1.55"
python-dotenv = "^0.17.1"
Jinja2 = "^2.11.3"
requests = "^2.25.1"
typer = "^0.3.2"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
